fix dtype mismatches for tf and torch backends in test_torch_gaussian_nll_loss where inputs of bfloat16 were being returned as float32. and increase tolerance to accommodate for mismatches like -0.3465735912322998!=-0.345703125

+++ ivy/functional/frontends/torch/nn/functional/loss_functions.py
@@ -368,7 +368,7 @@ def gaussian_nll_loss(input, target, var, full=False, eps=1e-6, reduction="mean"
     reduction = _get_reduction_func(reduction)
     ret = reduction(loss)
 
-    return ret
+    return ret.astype(input.dtype)
 
 
 @to_ivy_arrays_and_back
